requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.8
python-dotenv==1.0.0
django-filter==23.5
elasticsearch-dsl==8.13.1
